dbgapi: Add -Wconversion and fix whole dbgapi codebase
I tried to compile dbgapi with MSVC, and by default, MSVC issues a
large number of warnings around conversions, like:

 D:\therock\src\rocm-systems\projects\rocdbgapi\src\memory.h(191): warning C4244: 'argument': conversion from 'amd_dbgapi_size_t' to 'int', possible loss of data
 D:\therock\src\rocm-systems\projects\rocdbgapi\src\architecture.cpp(129): warning C4244: 'initializing': conversion from 'Integral' to 'const uint32_t', possible loss of data

... and many more.

I think that it's actually good to warn/error on implicit narrowing
conversions, which indeed can lose data unintentially.  Particularly
so since our codebase does a lot of bit manipulation.

So I went ahead and fixed all of the warning MSVC threw.  More
specifically, all warnings under /wd4244 and /wd4267:

  # warning C4244: '=': conversion from '...' to '...', possible loss of data
  # warning C4267: '=': conversion from '...' to '...', possible loss of data

I then went back to GCC, added -Wconversion, which warns for the same
things more or less, and fixed the remaining warnings, both on Linux
and mingw.

Then I did the same with Clang, both Linux and Windows.

This patch is the result.

Notes:

Instead of just static_cast, I added a new utils::narrow template function:

  template <typename To, typename From> To narrow (From from);

This is a safer version than static_cast -- it checks if we'd lose
data, before doing a static_cast.

The typical style is to use it like so when initializing a new
variable:

  auto foo = utils::narrow<uint32_t> (wider_type_value);

When assigning to a pre-existing variable, to avoid having to type the
type of the lhs variable, we have:

  template <typename To, typename From> void narrow_assign (To &to, From from);

Used like so:

  utils::narrow_assign (dest, src);

utils::narrow depends on std::numeric_limits to do its checking, and
because utils::narrow is passed {agent,global,host}_address_t, we need
to specialize std::numeric_limits for those types.

I added a new utils::split function that makes it convenient to split
a 64-bit value in two 32-bit parts, like so:

  auto [lo32, hi32] = utils::split (val64);

Many utils.h template functions hardcode parameters as int type.  I
made them template parameters too, which avoids having to cast at the
caller site.

bit_extract is tweaked to allow different source type and return type.

The following operators currently take an int as rhs, as that is a
distance type:

 constexpr amdgpu_regnum_t operator+ (amdgpu_regnum_t lhs, int rhs);
 constexpr amdgpu_regnum_t operator- (amdgpu_regnum_t lhs, int rhs);

Replace the "int" with a new amdgpu_regdiff_t typedef instead (which
is the signed counterpart of the underlying type of amdgpu_regnum_t
(which is uint32_t), which in practice means it's int32_t, so no real
type change.

However, the "operator-(amdgpu_regnum_t lhs, amdgpu_regnum_t rhs)"
function now returns amdgpu_regdiff_t, a signed integer, instead of
size_t, so that the result can naturally be passed directly as rhs to:

 constexpr amdgpu_regnum_t operator+ (amdgpu_regnum_t lhs, amdgpu_regdiff_t rhs);
 constexpr amdgpu_regnum_t operator- (amdgpu_regnum_t lhs, amdgpu_regdiff_t rhs);

Some local variables in architecture.cpp have amdgpu_regdiff_t though
at first look you'd think they should be size_t.  That's because if
you look closer to the code that actually uses them, you'll notice
that those variables exist only to be used as distances/offsets from
another register.

That's it.  Most of the patch is then passing explicit type to
utils::bit_mask, and utils::bit_extract, and using utils::narrow and
utils::narrow_assign.

commit-id:d8bac744

+++ src/utils.cpp
@@ -231,7 +231,7 @@ string_vprintf (const char *format, va_list va)
 
   dbgapi_assert (size >= 0);
 
-  std::string str (size, '\0');
+  std::string str (static_cast<size_t> (size), '\0');
   vsnprintf (&str[0], str.size () + 1, format, va);
   return str;
 }
